Latest development patch (#14)
* Update map.tscn

Added some Tile sources based on the sprites. Use this for map-based changes.

* Updated Project Settings + added tilemap

changed the texture filter to work with the sprites for Sprout Land and made the display more zoomed in when running the game

added tilemap (baseground, topground)

* added initial tiles

added tiles from sproutlands, to be changed

* lowered index of tiles

updated for the player to show up above the tiles

* Added School Sprites

added new sprite pack that includes school assets for map creation

* Imported the sprites inside TileMap

imported school sprites to use in tilemap and tilesets

* Updated and replaced some map tiles

Removed the sproutland tilesets in the map and made an initial classroom design map

* Made the Map Bigger and added objects

- added more space for future adjustments when needed
- added more objects for collision

* Added OutsideWalls Layer

for adding outside walls beyond the map

* Adding new game map into developmentBranch (#2)

* Update map.tscn

Added some Tile sources based on the sprites. Use this for map-based changes.

* Updated Project Settings + added tilemap

changed the texture filter to work with the sprites for Sprout Land and made the display more zoomed in when running the game

added tilemap (baseground, topground)

* added initial tiles

added tiles from sproutlands, to be changed

* lowered index of tiles

updated for the player to show up above the tiles

* Added School Sprites

added new sprite pack that includes school assets for map creation

* Imported the sprites inside TileMap

imported school sprites to use in tilemap and tilesets

* Updated and replaced some map tiles

Removed the sproutland tilesets in the map and made an initial classroom design map

* Made the Map Bigger and added objects

- added more space for future adjustments when needed
- added more objects for collision

* Added OutsideWalls Layer

for adding outside walls beyond the map

* Added Maze Room

added maze room below the normal classroom

* Delete map.tscn

* Update game_map.tscn

---------

Co-authored-by: Paolo <Paolo.encisa@obf.ateneo.edu>
Co-authored-by: Paolo <118610859+Pencisas@users.noreply.github.com>

* Revert "Adding new game map into developmentBranch (#2)"

This reverts commit ddab52772314cbf242f9edb517c3591c71696e9d.

* Added Maze Room and Bed Room

- added 3 layers in tilemap (mazeobjects: low, mid, and high)
- 2 rooms, maze and bed

* adding Map obstacles (#5)

* Added classmate ai

* update tscns

---------

Co-authored-by: Angelico Teaño <angelico.teano@obf.ateneo.edu>

* Changed some scale values and position

* Added Camera

Camera that follows the player.
TODO:
**Adjust values soon. Make it so the camera does NOT exceed the world map.

* Change child nodes to child scenes and change classmate size

* Add player animation and Fix camera limits

* Fixed player movement

Made it GRID BASED, and Stops on the middle of the tile.

* Add character obstacles and player idle animations (#7)

Co-authored-by: Angelico Teaño <angelico.teano@obf.ateneo.edu>

* Add character obstacles and player idle animations (#7) (#8)

Co-authored-by: Angelico Teaño <angelico.teano@obf.ateneo.edu>

* Grid-based movement patch (#9)

* Fixed player movement

Made it GRID BASED, and Stops on the middle of the tile.

* Add character obstacles and player idle animations (#7) (#8)

Co-authored-by: Angelico Teaño <angelico.teano@obf.ateneo.edu>

---------

Co-authored-by: Angelico Teaño <angelico.teano@obf.ateneo.edu>

* Update game_map.tscn

Fixed some map areas.

* Update game_map.tscn

---------

Co-authored-by: Paolo <Paolo.encisa@obf.ateneo.edu>
Co-authored-by: Paolo <118610859+Pencisas@users.noreply.github.com>
Co-authored-by: Angelico Teaño <angelico.teano@obf.ateneo.edu>